requirements.txt
flask==3.0.2
camel-ai==0.2.67 